+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to
 * Added a way to disable raw frame support: define the new
   `QNETHERNET_DISABLE_RAW_FRAME_SUPPORT` macro.
 * Added a "Complete list of features" section to the README.
+* Added `MDNSClass::hostname()` for the returning the hostname of the responder,
+  if running.
 
 ### Changed
 * Improved error code messages in `lwip_strerr(err)`. This is used when